Compile errors are fixed now to demo.
Compile errors for particle generators are fixed now.
It is now time to do a simple demo app for demostranting
the purpose of particle generators.
The main thing we had introduced was transforming the book
to trait based programming. Hopefully it would be easier to
extend the whole thing this way