A new camera and vector implementation, new camera needs stuff
The new camera implementation tries to be simpler but it
is not yet so simple.
It can advance in x, y, z direction in world space.
However those directions need to be fixed for camera space
that is we have to assume the position of camera as origin.
In order to properly handle this onb object is necessary.
It is already written but it seems to be buggy.

The next step is to handle smooth camera movement.
In fact it is imaginable to continue like this, if we
can get use to it. However it would be simpler and more proper
to fix it.

Once the camera issue is done, we need to work with setting up a
platform, like in the ballistic demo. This means basically that
the quat that we are passing to the scene has to have a larger
scale, and needs to be rotated along z axis.